+++ internal/common/advertiser.go
@@ -8,6 +8,6 @@ type Advertiser struct {
 	Name   string `json:"name,omitempty"`
 	Status string `json:"status,omitempty"`
 
-	ExchangeFee float32 `json:"exchangeFee,omitempty"` // Percentage (decimal)
-	DspFee      float32 `json:"dspFee,omitempty"`      // Percentage (decimal)
+	ExchangeFee float64 `json:"exchangeFee,omitempty"` // Percentage (decimal)
+	DspFee      float64 `json:"dspFee,omitempty"`      // Percentage (decimal)
 }
